Move client manager on fos bridge
