+++ psalm.xml
@@ -9,6 +9,7 @@
         <directory name="src" />
         <ignoreFiles>
             <directory name="vendor" />
+            <directory name="src/Monofony/Bridge/FOS/OAuthServer/spec" />
             <directory name="src/Monofony/Bridge/Sylius/User/spec" />
             <directory name="src/Monofony/Bundle/AdminBundle/Recipe/spec" />
             <directory name="src/Monofony/Bundle/AdminBundle/Recipe/tests" />
